exporting just the needed methods and macros
